When snapshotting, include the system library loc
This prevents errors in the following scenario:

- Bootstrap a package
- Add a dependency that has a "recommended" package as a dependency
  (for example, ggplot2 that depends on lattice)
- Try to snapshot

Without this commit, the snapshot fails because the lattice dependency
can't be found in lib.loc.

This commit isn't perfect in that the recommended package will now be
snapshotted (i.e. recorded in packrat.lock) but it won't actually be
in the private library. Will address that in a future commit.