Don't create source packages without root dirs
Source packages without a single root directory (i.e. with
./DESCRIPTION instead of ./$foo/DESCRIPTION, where $foo is
usually the package name but can also be packagename-sha1
or really anything at all) trigger a bug in
devtools::install_local_single where the contents of source
packages is unzipped directly to the tempdir and not cleaned
up correctly. When the second such package is installed in
the same R session, its file contents are merged with the
first package, which will usually lead to collation errors.

We'll also submit a patch to devtools but in the meantime
this fixes it and is probably what we want anyway.