use host provided build tools when available
The host of an R process (e.g. RStudio) may provide buildtools.check and buildtools.with options that can be used for automatic prompted installation of required build tools and execution of arbitrary code with the build tools on the PATH. This PR calls those optional functions whenever they are available.