Package isolation on Mac
We do a few things in this set of changes:
1. We symlink all 'base' and 'recommended' packages into a
'packrat/lib-R' library directory, and
2. Overwrite .Library to point at that directory.

By doing so, on Mac, we now have the packrat private library, and the
symlinked system library, which will now no longer pick up user
libraries that are installed there.

This would then obviate the need for library migration.