Split bundle into bundle_internal, bundle_external
When using R's internal tar, we're forced to copy the packrat project to
a temporary directory (because we cannot selectively filter out files
with the internal tar). However, this operation is expensive enough that
it warranted us having an implementation wrapping around R's internal
tar, and an unpenailzed external tar.