Fixes bug #288 - symlinking windows on networked drives
https://github.com/rstudio/packrat/issues/288

This is a windows problem. R/RStudio is on the C drive, and my project folder is on a networked drive. Windows will not allow a symlink/junction between the two drives, so instead of having a bunch of links, such as: packrat/lib-R/base I have a bunch of empty folders named: packrat/lib-R/base. Thus when "isPathToSamePackage" is called, packrat/lib-R/base/DESCRIPTION does not exist and an error happens. This just kills packrat, instead of returning a "FALSE" like it should.

Another option would be to edit the function "isPathToSamePackage" and include a "if(!file.exists(lhsPath) | file.exists(rhsPath)){ return(FALSE) }" before the "readChar" commands.