update install instructions to CRAN
As a sidenote, I just had a first run with profvis, the use of package::function is actually very costly in terms of time, I am not sure this so such a good practice. In this case, in the documentation of a package, I think it is safe to assume the user can load the package.